Add method to onboard and save new users to the db
Tested using local supabase instance

User_id is randomly generated and saved to db first.

If id is new, begin an onboarding process to create a userprofile. This userprofile is passed to LLM on each chat instance.

Created a separate onboarding API, and a separate userprofile API. Now Userdata can be queried from the home page using this API. Onboarding is now started automatically through this API on first time login.

LLM is now called through a util. This simiplified the onboard API and the regular chat API.

Visuazliation was removed for now in place of a better option